Add files via upload
Master Thesis project on Fairness in Geometry Processing. The thesis objective is to work on geometric methods that favor a representation of all human bodies in their diversity. We want to reach a mathematical representation that could put us one step ahead into the direction of more inclusive algorithms in AI.